feat: 일정 단건 조회 시, 해당 일정에 작성도니 모든 댓글 목록 조회 구현

+++ src/main/java/com/example/scheduledevelop2/comment/repository/CommentRepository.java
@@ -3,12 +3,13 @@ package com.example.scheduledevelop2.comment.repository;
 import com.example.scheduledevelop2.comment.entity.Comment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import java.util.List;
+
 public interface CommentRepository extends JpaRepository<Comment, Long> {
 
     default Comment findByIdOrElseThrow(Long commentId) {
         return findById(commentId).orElseThrow(()-> new IllegalArgumentException("존재하지 않는 댓글입니다.")); //todo: 예외 바꾸기
     }
 
-   int countBySchedule_Id(Long ScheduleId);
-
+    Long countBySchedule_Id(Long ScheduleId);
 }
